% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brute_force_knapsack.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{Knapsack solution - Brute Force}
\usage{
brute_force_knapsack(x, W, parallel = FALSE)
}
\arguments{
\item{x}{A DataFrame with two columns (weight(w) and value(v))}

\item{W}{Knapsack weight limit}

\item{parallel}{default = FALSE}
}
\value{
returns item with maxValue contained in knapsack
}
\description{
Knapsack solution - Brute Force
}
\references{
\url{https://en.wikipedia.org/wiki/Knapsack_problem#Dynamic_programming_in-advance_algorithm}
}
