% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_knapsack.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{Knapsack solution - Greedy}
\usage{
greedy_knapsack(x, W)
}
\arguments{
\item{x}{A DataFrame with two columns (weight(w) and value(v))}

\item{W}{Knapsack weight limit}
}
\value{
returns item with maxValue contained in knapsack
}
\description{
Knapsack solution - Greedy
}
\references{
\url{https://en.wikipedia.org/wiki/Knapsack_problem#Dynamic_programming_in-advance_algorithm}
}
