% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack_dynamic.R
\name{knapsack_dynamic}
\alias{knapsack_dynamic}
\title{Knapsack solution - Dynamic.}
\usage{
knapsack_dynamic(x, W)
}
\arguments{
\item{x}{A DataFrame with two columns (weight(w) and value(v))}

\item{W}{Knapsack weight limit}
}
\value{
returns item with maxValue contained in knapsack
}
\description{
Knapsack solution - Dynamic.
}
\references{
\url{https://en.wikipedia.org/wiki/Knapsack_problem#Dynamic_programming_in-advance_algorithm}
}
